added a default clock class that can be extended by other classes
