+++ src/plugins/remoteClock/RemoteClock.js
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Open MCT Web, Copyright (c) 2014-2015, United States Government
+ * Open MCT Web, Copyright (c) 2014-2021, United States Government
  * as represented by the Administrator of the National Aeronautics and Space
  * Administration. All rights reserved.
  *
@@ -19,100 +19,94 @@
  * this source code distribution or the Licensing information page available
  * at runtime from the About dialog for additional information.
  *****************************************************************************/
+import DefaultClock from '../../utils/clock/default-clock';
 
-define(['EventEmitter'], function (EventEmitter) {
-    /**
-     * A {@link openmct.TimeAPI.Clock} that updates the temporal bounds of the
-     * application based on UTC time values provided by a ticking local clock,
-     * with the periodicity specified.
-     * @param {number} period The periodicity with which the clock should tick
-     * @constructor
-     */
-    function LocalClock(period) {
-        EventEmitter.call(this);
+/**
+ * A {@link openmct.TimeAPI.Clock} that updates the temporal bounds of the
+ * application based on UTC time values provided by a ticking local clock,
+ * with the periodicity specified.
+ * @param {number} period The periodicity with which the clock should tick
+ * @constructor
+ */
 
-        /*
-        Metadata fields
-         */
+export default class RemoteClock extends DefaultClock {
+    constructor(openmct, identifier) {
+        super(openmct);
+
         this.key = 'remote-clock';
-        this.cssClass = 'icon-clock';
-        this.name = 'Remote Clock';
-        this.description = "Provides telemetry based timestamps from a configurable source.";
 
-        this.period = period;
-        this.timeoutHandle = undefined;
-        this.lastTick = Date.now();
+        // this.openmct = openmct;
+        // this.identifier = identifier;
+
+        // this.cssClass = 'icon-clock';
+        // this.name = 'Remote Clock';
+        // this.description = "Provides telemetry based timestamps from a configurable source.";
+
+        // // for now
+        // this.period = period;
+        // this.timeoutHandle = undefined;
+        // this.lastTick = Date.now();
+
+        console.log('remote clock installed for identifier', identifier);
     }
 
-    LocalClock.prototype = Object.create(EventEmitter.prototype);
+    // start() {
+    //     this.timeoutHandle = setTimeout(this.tick.bind(this), this.period);
+    // }
 
-    /**
-     * @private
-     */
-    LocalClock.prototype.start = function () {
-        this.timeoutHandle = setTimeout(this.tick.bind(this), this.period);
-    };
+    // stop() {
+    //     if (this.timeoutHandle) {
+    //         clearTimeout(this.timeoutHandle);
+    //         this.timeoutHandle = undefined;
+    //     }
+    // }
 
-    /**
-     * @private
-     */
-    LocalClock.prototype.stop = function () {
-        if (this.timeoutHandle) {
-            clearTimeout(this.timeoutHandle);
-            this.timeoutHandle = undefined;
-        }
-    };
+    // tick() {
+    //     const now = Date.now();
+    //     this.emit("tick", now);
+    //     this.lastTick = now;
+    //     this.timeoutHandle = setTimeout(this.tick.bind(this), this.period);
+    // }
 
-    /**
-     * @private
-     */
-    LocalClock.prototype.tick = function () {
-        const now = Date.now();
-        this.emit("tick", now);
-        this.lastTick = now;
-        this.timeoutHandle = setTimeout(this.tick.bind(this), this.period);
-    };
+    // /**
+    //  * Register a listener for the remote clock. When it ticks, the remote
+    //  * clock will provide the time from the configured endpoint
+    //  *
+    //  * @param listener
+    //  * @returns {function} a function for deregistering the provided listener
+    //  */
+    // on(event) {
+    //     const result = EventEmitter.prototype.on.apply(this, arguments);
 
-    /**
-     * Register a listener for the local clock. When it ticks, the local
-     * clock will provide the current local system time
-     *
-     * @param listener
-     * @returns {function} a function for deregistering the provided listener
-     */
-    LocalClock.prototype.on = function (event) {
-        const result = EventEmitter.prototype.on.apply(this, arguments);
+    //     if (this.listeners(event).length === 1) {
+    //         this.start();
+    //     }
 
-        if (this.listeners(event).length === 1) {
-            this.start();
-        }
+    //     return result;
+    // }
 
-        return result;
-    };
+    // /**
+    //  * Register a listener for the local clock. When it ticks, the local
+    //  * clock will provide the current local system time
+    //  *
+    //  * @param listener
+    //  * @returns {function} a function for deregistering the provided listener
+    //  */
+    // off(event) {
+    //     const result = EventEmitter.prototype.off.apply(this, arguments);
 
-    /**
-     * Register a listener for the local clock. When it ticks, the local
-     * clock will provide the current local system time
-     *
-     * @param listener
-     * @returns {function} a function for deregistering the provided listener
-     */
-    LocalClock.prototype.off = function (event) {
-        const result = EventEmitter.prototype.off.apply(this, arguments);
+    //     if (this.listeners(event).length === 0) {
+    //         this.stop();
+    //     }
 
-        if (this.listeners(event).length === 0) {
-            this.stop();
-        }
+    //     return result;
+    // }
 
-        return result;
-    };
+    // /**
+    //  * @returns {number} The last value provided for a clock tick
+    //  */
+    // currentValue() {
+    //     return this.lastTick;
+    // }
 
-    /**
-     * @returns {number} The last value provided for a clock tick
-     */
-    LocalClock.prototype.currentValue = function () {
-        return this.lastTick;
-    };
-
-    return LocalClock;
-});
+}
